Mahendra_Kumar_Reddy_Kakarla/NutritionFactChecker/requirements.txt
google-generativeai
pyyaml
Pillow
python-dotenv